--- apps/builder/app/shared/nano-states/folders.ts
@@ -1,4 +1,0 @@
-import { atom } from "nanostores";
-import type { Folders } from "@webstudio-is/sdk";
-
-export const $folders = atom<Folders>(new Map());
