+++ packages/project-build/src/index.ts
@@ -1,3 +1,2 @@
 export * from "./types";
 export * from "./shared/pages-utils";
-export * from "./shared/folder-utils";
