+++ packages/project-build/src/shared/pages-utils.ts
@@ -1,4 +1,4 @@
-import type { Pages, Page } from "@webstudio-is/sdk";
+import type { Pages, Page, Folder } from "@webstudio-is/sdk";
 
 export const findPageByIdOrPath = (
   pages: Pages,
@@ -10,4 +10,14 @@ export const findPageByIdOrPath = (
   return pages.pages.find(
     (page) => page.path === idOrPath || page.id === idOrPath
   );
+};
+
+// This is a root folder that nobody can delete or going to be able to see.
+export const createRootFolder = (children: Folder["children"] = []): Folder => {
+  return {
+    id: "root",
+    name: "Root",
+    slug: "",
+    children,
+  };
 };
