+++ packages/sdk/src/schema/pages.ts
@@ -2,7 +2,7 @@ import { z } from "zod";
 
 const MIN_TITLE_LENGTH = 2;
 
-export const PageId = z.string();
+const PageId = z.string();
 
 export const PageName = z
   .string()
@@ -103,3 +103,31 @@ export const Pages = z.object({
 });
 
 export type Pages = z.infer<typeof Pages>;
+
+const FolderId = z.string();
+
+export const FolderName = z
+  .string()
+  .refine((value) => value.trim() !== "", "Can't be empty");
+
+export const FolderSlug = z
+  .string()
+  .refine((slug) => slug !== "", "Can't be empty")
+  .refine((slug) => slug.includes("/") === false, "Can't contain a /")
+  .refine(
+    (path) => /^[-a-z0-9]*$/.test(path),
+    "Only a-z, 0-9 and - are allowed"
+  );
+
+export const Folder = z.object({
+  id: FolderId,
+  name: FolderName,
+  slug: z.string(),
+  children: z.array(z.union([FolderId, PageId])),
+});
+
+export type Folder = z.infer<typeof Folder>;
+
+export const Folders = z.map(FolderId, Folder);
+
+export type Folders = z.infer<typeof Folders>;
